pyproject.toml
[tool.poetry]
name = "school-diary"
version = "2.0"
description = "Open-source digital diary for schools made with Django."
authors = ["AlanTheKnight <ideasoft-spb@yandex.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-filter = "^2.4.0"
django-cleanup = "^5.1.0"
django-extensions = "^3.1.2"
djangorestframework = "^3.12.4"
django-widget-tweaks = "^1.4.8"
Markdown = "^3.3.4"
Pillow = "^8.2.0"
django-debug-toolbar = "^3.2"
XlsxWriter = "^1.3.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
